refactor: dataset textStyle
