+++ opts/global.go
@@ -235,31 +235,6 @@ type AxisLine struct {
 	LineStyle *LineStyle `json:"lineStyle,omitempty"`
 }
 
-// TextStyle is the option set for a text style component.
-type TextStyle struct {
-	// Font color
-	Color string `json:"color,omitempty"`
-
-	// Font style
-	// Options: 'normal', 'italic', 'oblique'
-	FontStyle string `json:"fontStyle,omitempty"`
-
-	// Font size
-	FontSize int `json:"fontSize,omitempty"`
-
-	// Font family the main title font family.
-	// Options: "sans-serif", 'serif' , 'monospace', 'Arial', 'Courier New', 'Microsoft YaHei', ...
-	FontFamily string `json:"fontFamily,omitempty"`
-
-	// Padding title space around content.
-	// The unit is px. Default values for each position are 5.
-	// And they can be set to different values with left, right, top, and bottom.
-	Padding interface{} `json:"padding,omitempty"`
-
-	// compatible for WordCloud
-	Normal *TextStyle `json:"normal,omitempty"`
-}
-
 // SplitArea is the option set for a split area.
 type SplitArea struct {
 	// Set this to true to show the splitArea.
@@ -567,12 +542,4 @@ type ViewControl struct {
 	// Rotate Speed, (angle/s).
 	// default 10
 	AutoRotateSpeed float32 `json:"autoRotateSpeed,omitempty"`
-}
-
-// Dataset brings convenience in data management separated with styles and enables data reuse by different series.
-// More importantly, it enables data encoding from data to visual, which brings convenience in some scenarios.
-// https://echarts.apache.org/en/option.html#dataset.id
-type Dataset struct {
-	//source
-	Source interface{} `json:"source"`
 }
