+++ opts/text_style.go
@@ -1,0 +1,26 @@
+package opts
+
+// TextStyle is the option set for a text style component.
+type TextStyle struct {
+	// Font color
+	Color string `json:"color,omitempty"`
+
+	// Font style
+	// Options: 'normal', 'italic', 'oblique'
+	FontStyle string `json:"fontStyle,omitempty"`
+
+	// Font size
+	FontSize int `json:"fontSize,omitempty"`
+
+	// Font family the main title font family.
+	// Options: "sans-serif", 'serif' , 'monospace', 'Arial', 'Courier New', 'Microsoft YaHei', ...
+	FontFamily string `json:"fontFamily,omitempty"`
+
+	// Padding title space around content. See legend.textStyle.padding
+	// The unit is px. Default values for each position are 5.
+	// And they can be set to different values with left, right, top, and bottom.
+	Padding interface{} `json:"padding,omitempty"`
+
+	// compatible for WordCloud
+	Normal *TextStyle `json:"normal,omitempty"`
+}
